third_party/WebKit/LayoutTests/fast/css/nth-child-dynamic-002-expected.txt

An existing LocalStyleChange on parent should not inhibit forward/backward rule update on children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(inner).color is "rgb(255, 255, 255)"
PASS getComputedStyle(inner).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE


You should see a green square below, and no FAIL.

FAIL
